+++ motiiv-front/src/components/common/Login/FirstPage.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
-import btnNaver from '../../../../assets/global/btn_naver.png';
-import btnKakao from '../../../../assets/global/btn_kakao.png';
+import btnNaver from '../../../assets/global/btn_naver.png';
+import btnKakao from '../../../assets/global/btn_kakao.png';
 
 const Container = styled.div`
   display: ${props => props.page === 1 ? 'flex' : 'none' };
