+++ motiiv-front/src/components/common/Navbar/Navbar.js
@@ -5,8 +5,8 @@ import { useState } from 'react';
 import { useSelector } from 'react-redux';
 import logo from '../../../assets/global/motiiv_logo.png';
 import star from '../../../assets/global/star.png';
-import SigninModal from './SigninModal/SignInModal';
-import ProfileModal from './ProfileModal';
+import SigninModal from '../Login/SignInModal';
+import ProfileModal from './profilemodal/ProfileModal';
 
 const activeStyle = {
   color: '#2cff2c',
@@ -47,6 +47,17 @@ const TabElem = styled(NavLink)`
   color: white;
   text-decoration: none;
   padding: 1.6rem 0.8rem;
+
+  ${props => props.hightlight == true? 
+    `
+      color: #2cff2c;
+      border-bottom: 0.2rem solid #2cff2c;
+    `
+    :
+    `
+
+    `}
+
   &:nth-child(7) {
     display: ${props => (props.show === 'true' ? 'flex' : 'none')};
   }
@@ -72,7 +83,7 @@ const LoginContainer = styled.div`
 `;
 
 const Login = styled.div`
-  display: ${props => (props.login === 'false' ? 'flex' : 'none')};
+  display: ${props => (props.login === false ? 'flex' : 'none')};
   color: white;
   text-align: left;
   text-decoration: none;
@@ -83,10 +94,19 @@ const FirstLetter = styled.div`
   color: #2cff2c;
   font-size: 1.75rem;
   font-family: 'Spoqa-Han-Sans';
+
+  ${props => props.isShow ?
+  `
+    display:none;
+  `
+  :
+  `
+  `
+  };
 `;
 
 const Profile = styled.div`
-  display: ${props => (props.login === 'true' ? 'flex' : 'none')};
+  display: ${props => (props.login === true ? 'flex' : 'none')};
   width: 3rem;
   height: 3rem;
   z-index: 3;
@@ -118,12 +138,14 @@ const Profile = styled.div`
   }
 `;
 
-function Navbar() {
-  //로그인 여부 판단 + 어드민 여부 판단
-  const [loginState, setLoginState] = useState({
-    isLoggined: true,
-    admin: false,
-  });
+function Navbar({location, isloggined}) {
+
+  const { userInfo } = useSelector(({ user }) => ({
+    userInfo: user.userInfo
+  }));
+  console.log(userInfo);
+
+  const pathname = location.substr(1, 1);
 
   //프로필 드롭다운 나타나고 없애기
   const [profileModalState, setProfileModalState] = useState(false);
@@ -145,65 +167,63 @@ function Navbar() {
     })();
   };
 
-  const { userInfo } = useSelector(({ user }) => ({
-    userInfo: user.userInfo,
-  }));
-
-  const name = 'Bonnie';
+  const name = "어쩌구"
   const firstletter = name.substr(0, 1);
   //첫글자가 영어인지 한글인지 테스트하는 로직 필요
 
   return (
     <>
       <Header>
+        
         <NavLink exact to="/main">
           <Logo src={logo} />
         </NavLink>
+
         <TabContainer>
-          <TabElem exact to="/main" activeStyle={activeStyle}>
+          <TabElem
+            exact to="/main"
+            activeStyle={activeStyle}
+          >
             main
           </TabElem>
           <Star src={star} />
-          <TabElem to="/category/0" activeStyle={activeStyle}>
+          <TabElem
+            to="/category/0"
+            activeStyle={activeStyle}
+          >
             category
           </TabElem>
           <Star src={star} />
-          <TabElem exact to="/mymotiiv" activeStyle={activeStyle}>
-            mymotiiv
-          </TabElem>
-          <Star src={star} show={loginState.admin.toString()} />
           <TabElem
-            exact
-            to="/admin"
-            show={loginState.admin.toString()}
+            exact to="/mymotiiv"
             activeStyle={activeStyle}
           >
-            admin
+            mymotiiv
           </TabElem>
         </TabContainer>
 
         <LoginContainer>
           <Login
-            login={loginState.isLoggined.toString()}
+            login={isloggined.isLogin}
             onClick={onClickLoginBtn}
           >
             login
           </Login>
           <Profile
             src={userInfo.profileImageUrl}
-            login={loginState.isLoggined.toString()}
+            login={isloggined.isLogin}
             onClick={onClickProfileImage}
             onclick={profileModalState}
-          >
-            <FirstLetter>{firstletter}</FirstLetter>
+          ><FirstLetter isShow = {userInfo.profileImageUrl}>{firstletter}</FirstLetter>
           </Profile>
           <ProfileModal
             showModal={profileModalState}
-            name={name}
+            name={userInfo.username}
             firstletter={firstletter}
           />
         </LoginContainer>
       </Header>
+
       <SigninModal showModal={loginModalState} />
     </>
   );
