+++ motiiv-front/src/lib/api/_client.js
@@ -5,7 +5,8 @@ const client = axios.create({
   headers: {
     'Content-Type': 'application/json',
     credentials: true,
-    //jwt:'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MTQsInVzZXJuYW1lIjoi7Jqw7JiBIiwic25zSWQiOiIxIiwic29jaWFsVHlwZSI6Imtha2FvIiwiaWF0IjoxNjEwMDA4NzU1LCJleHAiOjE2MTI2MDA3NTUsImlzcyI6Im1vdGlpdiJ9.1_CTWbQWps8vcpPUzzKuAYFy7Th4fqpIPummhhabEm8',
+    userToken:
+      'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MTQsInVzZXJuYW1lIjoi7Jqw7JiBIiwic25zSWQiOiIxIiwic29jaWFsVHlwZSI6Imtha2FvIiwiaWF0IjoxNjEwMjU3Mzk5LCJleHAiOjE2MTI4NDkzOTksImlzcyI6Im1vdGlpdiJ9.3nwQuSC8xJEX7oPr0IFNxJoe45YosvfXT953c4Vws_s',
   },
 });
 //ApiConfig.defaults.headers.jwt = data.data.accessToken; // api 토큰 바꿔치기
